npm:d3 -> dt:d3 and added dt:dagre
dt:dagre-d3 refers to ../d3/d3.d.ts, which exists in the dt repo but not
in the npm repo. Not to mention, the d3 typing definitions found in the
npm repo must be installed as modules, so dagre-d3 would have to
reference ../../modules/... to access it.

Dagre typings were simply just missing, but referenced by dagre-d3
typings.